+++ lib/mutex_write.go
